test(HeightTest): 원시 값 포장으로 인한 테스트 코드 수정
